requirements.txt
Django==4.2.2
pytz==2023.3
djangorestframework==3.14.0
django_rest_passwordreset==1.3.0
requests==2.31.0
ujson==5.8.0
pyYAML==6.0.1
python-dotenv==1.0.0
psycopg2==2.9.6
django-filter==23.2